% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_bigwig.R
\name{bam_to_bigwig}
\alias{bam_to_bigwig}
\title{Convert a BAM file to a BigWig}
\usage{
bam_to_bigwig(
  bam_file,
  prefix = file.path(tempdir(), paste0(basename(bam_file))),
  min_unique_qual = 10,
  double_count = FALSE
)
}
\arguments{
\item{bam_file}{A \code{character(1)} with the path to the input BAM file.}

\item{prefix}{A \code{character(1)} specifying the output file prefix. This
function creates a file called \code{prefix.all.bw} that can be read again later
with \code{read_coverage()}. By default, the prefix is the BAM file name and the
file is created in the \code{tempdir()} and will be deleted after you close your
R session.}

\item{min_unique_qual}{A \code{integer(1)} specifying a mapping quality threshold
and only bases above this will be used to generate the BigWig.}

\item{double_count}{A \code{logical(1)} determining whether to count the
overlapping ends of paired ends reads twice.}
}
\value{
A \code{character(1)} with the path to the generated BigWig file.
}
\description{
Given an input BAM file, convert this to the BigWig format which can then be
used in \code{get_coverage}.
}
\examples{

## Install if necessary
install_megadepth()

## Intialise path to the example BAM file
example_bam <- system.file("tests", "test.bam",
    package = "megadepth", mustWork = TRUE
)

example_bw <- bam_to_bigwig(example_bam)

## print path to generated BigWig
example_bw
}
