% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_megadepth}
\alias{install_megadepth}
\title{Install Megadepth}
\usage{
install_megadepth(version = "latest", force = FALSE)
}
\arguments{
\item{version}{A \verb{character(})` specifying the Megadepth version number,
e.g., \code{1.0.4}; the special value \code{latest} means the latest version
(fetched from Github releases).}

\item{force}{A \code{logical(1)} specifying whether to install megadepth even if
it has already been installed.}
}
\value{

}
\description{
Download the appropriate Megadepth executable for your platform from Github
and try to copy it to a system directory so \pkg{megadepth} can run the
\command{megadepth} command.
}
\examples{

\dontrun{
install_megadepth()
}
}
\references{
This function is based on blogdown::install_hugo() which is available from
\url{https://github.com/rstudio/blogdown/blob/master/R/install.R}.
}
