% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megadepth.R
\name{megadepth_cmd}
\alias{megadepth_cmd}
\title{Run Megadepth commands}
\usage{
megadepth_cmd(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{system2('megadepth', ...)}, e.g.
\code{annotation(path)} is basically
\code{megadepth_cmd(c('--annotation', path))}
(i.e. run the command \command{megadepth --annotation path}).}
}
\description{
Wrapper functions to run Megadepth commands via
\code{\link{system2}('megadepth', ...)}.
}
\section{Functions}{
\itemize{
\item \code{megadepth_cmd}: Run an arbitrary Megadepth command.
}}

\examples{

## Install if necessary
install_megadepth()

## Find version
megadepth_cmd("--version", stdout = TRUE)
}
\references{
This function is based on blogdown::hugo_cmd() which is available from
\url{https://github.com/rstudio/blogdown/blob/master/R/hugo.R}.
}
