% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coverage.R
\name{read_coverage}
\alias{read_coverage}
\title{Read a coverage TSV file created by Megadepth}
\usage{
read_coverage(tsv_file, verbose = TRUE)
}
\arguments{
\item{tsv_file}{A \code{character(1)} specifying the path to the tab-separated
(TSV) file created manually using \code{megadepth_shell()} or on a previous
\code{get_coverage()} run.}

\item{verbose}{A \code{logical(1)} controlling whether to suppress messages when
reading the data.}
}
\value{
A \link[GenomicRanges:GRanges-class]{GRanges-class} object with the
coverage summarization across the annotation ranges.
}
\description{
Read an \verb{*annotation.tsv} file created by \code{get_coverage()} or manually by
the user using Megadepth.
}
\examples{

## For RleList, equivalent to rtracklayer::import.bw(as = "RleList")
## GenomicRanges::coverage(x, weights = "cov")
}
\seealso{
Other Coverage functions: 
\code{\link{get_coverage}()}
}
\concept{Coverage functions}
