% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_junction.R
\name{read_junction_table}
\alias{read_junction_table}
\title{Read a junction TSV file created by Megadepth as a table}
\usage{
read_junction_table(tsv_file)
}
\arguments{
\item{tsv_file}{A \code{character(1)} specifying the path to the tab-separated
(TSV) file created manually using \code{megadepth_shell()} or on a previous
\code{bam_to_junctions()} run.}
}
\value{
A \code{tibble::tibble()} with the junction data that follows the format
specified at
\url{https://github.com/ChristopherWilks/megadepth#megadepth-pathtobamfile---junctions}.
}
\description{
Read an \verb{*jxs.tsv} file created by \code{bam_to_junctions()} or manually by
the user using Megadepth.
}
\examples{

## Install if necessary
install_megadepth()

## Find the example BAM file
example_bam <- system.file("tests", "test.bam",
    package = "megadepth", mustWork = TRUE
)

## Run bam_to_junctions()
example_jxs <- bam_to_junctions(example_bam, overwrite = TRUE)

# ## Read the data as a tibble using the format specified at
# ## https://github.com/ChristopherWilks/megadepth#megadepth-pathtobamfile---junctions
# read_junction_table(example_jxs)
}
