% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_junctions.R
\name{bam_to_junctions}
\alias{bam_to_junctions}
\title{Extract junctions from a BAM file}
\usage{
bam_to_junctions(
  bam_file,
  prefix = file.path(tempdir(), paste0(basename(bam_file))),
  long_reads = FALSE
)
}
\arguments{
\item{bam_file}{A \code{character(1)} with the path to the input BAM file.}

\item{prefix}{A \code{character(1)} specifying the output file prefix. This
function creates a file called \code{prefix.jxs.tsv}. By default, the prefix is
the BAM file name and the file is created in the \code{tempdir()} and will be
deleted after you close your R session.}

\item{long_reads}{A \code{logical(1)} indicating whether to increase the buffer
size to accomodate for long-read RNA-sequencing.}
}
\value{
A \code{character(1)} with the path to the generated junctions TSV file.
}
\description{
Given a BAM file, extract junction information including co-ordinates,
strand, anchor length for each junction read.
}
\examples{

## Install if necessary
install_megadepth()

if (!exists("example_bam")) {
    ## Intialise path to the example BAM file
    example_bam <- system.file("tests", "test.bam",
        package = "megadepth", mustWork = TRUE
    )
}

example_jxs <- bam_to_junctions(example_bam)

## print path to generated BigWig
example_jxs
}
